option/stats: merge to trx
