--- src/tr1/game/option/stats.h
@@ -1,3 +1,0 @@
-#pragma once
-
-#include <libtrx/game/option/stats.h>
