+++ src/tr2/meson.build
@@ -100,7 +100,6 @@ sources = [
   'game/objects/general/harpoon_bolt.c',
   'game/objects/general/zipline.c',
   'game/objects/setup.c',
-  'game/option/stats.c',
   'game/savegame/savegame_bson.c',
   'game/savegame/savegame_legacy.c',
   'game/stats.c',
